Splitting Sendable.convert_video() tests out from others
